Set Puppeteer module's launch option to headless
The launch option of the Puppeteer module in both the main app and crawler component has been set to 'headless: true'. This change will improve testing by running tests without a UI. This is beneficial when the tests are being run in an environment without a graphical display such as a continuous integration server.